Moved abstract component and common helper to common folder
