+++ game_app/common/abstract_component.py
