+++ game_app/common/common_helper.py
