+++ game_app/menus/credits/credits_logic.py
@@ -9,9 +9,9 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 
 from pygame.locals import *
 
-from game_app.common_helper import Components
+from game_app.common.common_helper import Components
 from game_app.common.buttons import RoundIconButton
-from game_app.common_helper import Settings, ColorMode
+from game_app.common.common_helper import Settings, ColorMode
 
 
 class CreditsLogic:
