+++ game_app/menus/credits/credits_scene.py
@@ -8,7 +8,7 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 # -------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
 
 import pygame
-from game_app.common_helper import Settings, ColorMode
+from game_app.common.common_helper import Settings, ColorMode
 
 
 class CreditsScene:
