+++ game_app/menus/tic_tac_toe_launch_menu/tic_tac_toe_launch_menu_logic.py
@@ -8,7 +8,7 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 # -------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
 
 from pygame.locals import *
-from game_app.common_helper import Components, ColorMode, Settings, Difficulty
+from game_app.common.common_helper import Components, ColorMode, Settings, Difficulty
 from game_app.common.buttons import RectangularChoiceButton, RectangularTextButton, RoundIconButton, \
     RectangularChoiceButtonWithSubtext
