+++ game_app/menus/tic_tac_toe_launch_menu/tic_tac_toe_launch_menu_scene.py
@@ -7,7 +7,7 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 #-------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
 
 import pygame
-from game_app.common_helper import ColorMode, Settings
+from game_app.common.common_helper import ColorMode, Settings
 
 
 class TicTacToeLaunchMenuScene:
